pyproject.toml
[tool.poetry]
name = "not-decided"
version = "0.1.0"
description = ""
authors = ["Rohit <rk22012001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
redis = "^5.2.1"
passlib = "^1.7.4"
psycopg2 = "^2.9.10"
bcrypt = "4.0.1"
pyjwt = "^2.10.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
ssort = "^0.14.0"
isort = "^5.13.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"